+++ src/modules/client/ClientProfile/components/AccountCard/AccountCard.module.scss
@@ -1,0 +1,46 @@
+.block {
+  display: flex;
+  align-items: center;
+  gap: 30px;
+  padding: 10px;
+  padding-right: 30px;
+  min-width: 440px;
+  width: max-content;
+  background-color: var(--light-gray);
+  cursor: pointer;
+
+  div {
+    display: flex;
+    flex-direction: column;
+  }
+
+  span {
+    font-size: 24px;
+  }
+
+  .avatar {
+    circle {
+      fill: var(--gray);
+    }
+  }
+
+  .logout-icon {
+    margin-left: auto;
+    width: 40px;
+    height: 40px;
+
+    .logout-arrow {
+      transition: transform 0.15s ease-in-out;
+    }
+
+    &:hover {
+      .logout-arrow {
+        transform: translateX(5px);
+      }
+    }
+  }
+}
+
+.active {
+  outline: 3px solid var(--red);
+}
