--- src/modules/client/ClientProfile/components/AccountsList/AccountsList.module.scss
@@ -1,52 +1,0 @@
-.accounts-list {
-  display: inline-flex;
-  gap: 20px;
-}
-
-.client {
-  width: 440px;
-  cursor: pointer;
-
-  .block {
-    display: flex;
-    align-items: center;
-    gap: 30px;
-    padding: 10px;
-    width: 100%;
-    background-color: var(--light-gray);
-
-    div {
-      display: flex;
-      flex-direction: column;
-    }
-
-    span {
-      font-size: 24px;
-    }
-
-    .avatar {
-      circle {
-        fill: var(--gray);
-      }
-    }
-
-    .logout-icon {
-      width: 40px;
-      height: 40px;
-
-      .logout-arrow {
-        transition: transform 0.15s ease-in-out;
-      }
-
-      &:hover {
-        .logout-arrow {
-          transform: translateX(5px);
-        }
-      }
-    }
-  }
-
-  .active {
-    outline: 3px solid var(--red);
-  }
-}
