+++ src/modules/client/ClientProfile/components/AccountsModal/index.tsx
@@ -1,17 +1,23 @@
 import { FC } from 'react';
 import { Modal } from 'antd';
 import { Button } from '@UI';
+import { User } from '@services/authApi';
+import { AccountCard } from '../AccountCard';
 
 interface AccountsModal {
   onCancel(): void;
   isOpen: boolean;
-  onLogout(): void;
+  onLogout(e: React.MouseEvent, id: number): void;
+  userProfiles: User[];
+  activeUserId: number;
 }
 
 export const AccountsModal: FC<AccountsModal> = ({
   onCancel,
   isOpen,
   onLogout,
+  userProfiles,
+  activeUserId,
 }) => {
   return (
     <Modal
@@ -20,12 +26,23 @@ export const AccountsModal: FC<AccountsModal> = ({
       open={isOpen}
       footer={
         <div style={{ display: 'flex', gap: 20 }}>
-          <Button onClick={onLogout} to="/admin-auth">
+          <Button onClick={(e) => onLogout(e, activeUserId)} to="/admin-auth">
             Сменить аккаунт
           </Button>
           <Button to="/admin-auth">Добавить аккаунт</Button>
         </div>
       }
-    />
+    >
+      <ul style={{ display: 'flex', flexDirection: 'column', gap: 20 }}>
+        {userProfiles.map((user) => (
+          <AccountCard
+            user={user}
+            activeUserId={activeUserId}
+            onLogout={onLogout}
+            onCancelModal={onCancel}
+          />
+        ))}
+      </ul>
+    </Modal>
   );
 };
